requirements.txt
requests
python-dotenv
tabulate
fastapi
uvicorn[standard]
websockets
